have a more organized front end that works with the backend. Now I have api that are doing great with their calls.
