+++ shift-scan/src/app/(content)/hoursControl.tsx
@@ -6,6 +6,8 @@ import { useTranslations } from "next-intl";
 import { Contents } from "@/components/(reusable)/contents";
 import { Texts } from "@/components/(reusable)/texts";
 import { Buttons } from "@/components/(reusable)/buttons";
+import { Holds } from "@/components/(reusable)/holds";
+import { Titles } from "@/components/(reusable)/titles";
 
 type ControlComponentProps = {
   toggle: (toggle: boolean) => void;
@@ -129,38 +131,32 @@ export default function ControlComponent({ toggle } : ControlComponentProps) {
   };
 
   return (
-    <>
-      <Contents
-        variant={"hoursDisplay"}
-        size={null}
-        title={t("DA-Control-Title")}
-      >
+    <Holds className="w-full h-fit">
         {/* Th */}
         <ViewComponent
           scrollLeft={scrollLeft}
           scrollRight={scrollRight}
           returnToMain={returnToMain}
           currentDate={currentDate}
-        />
+          />
         {/* This is the start of the bar chart componnet and the previous day */}
-        <Contents variant={"barChartWrapper"} size={null}>
+          <Holds className="flex flex-row justify-center rounded-2xl shadow-[8px_8px_0px_grey] p-4 border-4 border-black w-full h-fit bg-white mt-4">
           {/* Contexts gives base styles for background app-dark-blue bars */}
-          <Contents variant={"navy"} size={"defaultHours"}>
+          <Holds className="bg-app-dark-blue w-[100px] h-[300px] mx-auto rounded-2xl pt-3 pb-3 px-2 flex flex-col justify-end">
             {/* - Contexts gives base styles for green bar with dark-blue background
     it also provides the height of the green bar, based on the value and the color of the bar based on the value
     - If the value is less than 8 hours are orange, greater than or equal to 8 hours are green
     - Text Component is neccessary to give high to the content while having a blank space
-*/}
-            <Contents
-              variant={"default"}
-              size={null}
+    */}
+            <Holds
+             
               /*
-1. ternary is used for height evavulation based on the value, it uses caluclate bar to get it in px 
-2. ternary is used for color evavulation based on the value, it uses caluclate bar to get it in px
-3. ternary is used to show blank non working days as a clear background rather then show any status
-*/
-              className={`w-full rounded-2xl flex flex-col justify-end
-          ${
+              1. ternary is used for height evavulation based on the value, it uses caluclate bar to get it in px 
+              2. ternary is used for color evavulation based on the value, it uses caluclate bar to get it in px
+              3. ternary is used to show blank non working days as a clear background rather then show any status
+              */
+             className={`w-full rounded-2xl flex flex-col justify-end h-fit max-w-lg mx-auto pt-10
+              ${
             currentData.valuePrev === 0
               ? "bg-clear"
               : `h-[${calculateBarHeight(currentData.valuePrev)}px]`
@@ -174,62 +170,58 @@ export default function ControlComponent({ toggle } : ControlComponentProps) {
                   ? `${currentData.valuePrev.toFixed(1)} ${t("DA-Time-Label")}`
                   : ""}
               </Texts>
-            </Contents>
-          </Contents>
+            </Holds>
+            </Holds>
           {/* This is the current day bar same as the previous with styling */}
-          <Contents variant={"navy"} size={"defaultHours"}>
-            <Contents
-              variant={"default"}
-              size={null}
-              className={` w-full flex flex-col justify-end rounded-2xl 
-        ${
-          currentData.value === 0
-            ? "bg-clear"
-            : `h-[${calculateBarHeight(currentData.value)}px]`
-        }  
-        ${currentData.value > 8 ? "bg-app-green" : "bg-app-orange"}
-        ${currentData.value !== 0 ? "" : "bg-clear"}
-        `}
-            >
+          <Holds className="bg-app-dark-blue w-[100px] h-[300px] mx-auto rounded-2xl pt-3 pb-3 px-2 flex flex-col justify-end">
+            <Holds
+              className={` w-full flex flex-col justify-end rounded-2xl h-fit max-w-lg mx-auto pt-10
+                ${
+                  currentData.value === 0
+                  ? "bg-clear"
+                  : `h-[${calculateBarHeight(currentData.value)}px]`
+                  }  
+                  ${currentData.value > 8 ? "bg-app-green" : "bg-app-orange"}
+                  ${currentData.value !== 0 ? "" : "bg-clear"}
+                  `}
+                  >
               <Texts variant={"default"} size={"p3"}>
                 {currentData.value !== 0
                   ? `${currentData.value.toFixed(1)} ${t("DA-Time-Label")}`
                   : ""}
               </Texts>
-            </Contents>
-          </Contents>
+            </Holds>
+          </Holds>
+
+
           {/* This is the next day bar same as the previous with styling */}
-          <Contents variant={"navy"} size={"defaultHours"}>
-            <Contents
-              variant={"default"}
-              size={null}
-              className={` w-full flex flex-col justify-end rounded-2xl 
-        ${
-          currentData.valueNext === 0
-            ? "bg-clear"
-            : `h-[${calculateBarHeight(currentData.valueNext)}px] `
-        }
-        ${currentData.valueNext > 8 ? "bg-app-green" : "bg-app-orange"}
-        ${currentData.valueNext !== 0 ? "" : "bg-clear"}
-        `}
-            >
+          <Holds className="bg-app-dark-blue w-[100px] h-[300px] mx-auto rounded-2xl pt-3 pb-3 px-2 flex flex-col justify-end">
+            <Holds
+              className={`h-fit max-w-lg mx-auto pt-10  w-full flex flex-col justify-end rounded-2xl 
+                ${
+                  currentData.valueNext === 0
+                  ? "bg-clear"
+                  : `h-[${calculateBarHeight(currentData.valueNext)}px] `
+                  }
+                  ${currentData.valueNext > 8 ? "bg-app-green" : "bg-app-orange"}
+                  ${currentData.valueNext !== 0 ? "" : "bg-clear"}
+                  `}
+                  >
               <Texts variant={"default"} size={"p3"}>
                 {currentData.valueNext !== 0
                   ? `${currentData.valueNext.toFixed(1)} ${t("DA-Time-Label")}`
                   : ""}
               </Texts>
-            </Contents>
-          </Contents>
-        </Contents>
-        <Contents variant={null} size={null}>
+            </Holds>
+          </Holds>
+        </Holds>
+
           <Buttons href="/timesheets" variant={"green"} size={"fill"}>
             <Texts variant={"default"} size={"p3"}>
               View My Timesheets
             </Texts>
           </Buttons>
-        </Contents>
-      </Contents>
-    </>
+        </Holds>
   );
 };
