+++ shift-scan/src/app/context/dbRecentCodesContext.tsx
@@ -1,7 +1,7 @@
 // This stores the previous 5 cost codes, jobsites, and equipment that the user has selected. This will make it easier to change cost codes.
 
 import React, { createContext, useContext, useState, ReactNode } from 'react';
-import { JobCodes, CostCode, EquipmentCodes } from '@/lib/types';
+import { JobCodes, CostCodes, EquipmentCodes } from '@/lib/types';
 
 type RecentJobSiteContextType = {
   recentlyUsedJobCodes: JobCodes[];
@@ -36,9 +36,9 @@ export const useRecentDBJobsite = () => useContext(RecentJobSiteContext);
 
 
 interface RecentCostCodeContextType {
-    recentlyUsedCostCodes: CostCode[];
-    setRecentlyUsedCostCodes: React.Dispatch<React.SetStateAction<CostCode[]>>;
-    addRecentlyUsedCostCode: (code: CostCode) => void;
+    recentlyUsedCostCodes: CostCodes[];
+    setRecentlyUsedCostCodes: React.Dispatch<React.SetStateAction<CostCodes[]>>;
+    addRecentlyUsedCostCode: (code: CostCodes) => void;
   }
   
   const RecentCostCodeContext = createContext<RecentCostCodeContextType>({
@@ -48,9 +48,9 @@ interface RecentCostCodeContextType {
   });
   
   export const RecentCostCodeProvider = ({ children }: { children: ReactNode }) => {
-    const [recentlyUsedCostCodes, setRecentlyUsedCostCodes] = useState<CostCode[]>([]);
+    const [recentlyUsedCostCodes, setRecentlyUsedCostCodes] = useState<CostCodes[]>([]);
   
-    const addRecentlyUsedCostCode = (code: CostCode) => {
+    const addRecentlyUsedCostCode = (code: CostCodes) => {
       setRecentlyUsedCostCodes((prev) => {
         const updatedList = [code, ...prev.filter((c) => c.id !== code.id)];
         return updatedList.slice(0, 5);
